Add automation for converting DAT files
All you have to do is drop the dats in the correct folder and it'll convert everything for you automatically. It gives you a parkobj file where you just have to edit object.json and the name. Still a bit of work, but should take away alot of the hand work.